+++ config/me_cms.php
@@ -80,6 +80,17 @@ return ['MeCms' => [
 		//Site title
 		'title' => 'MeCms'
 	],
+	//Pages
+	'page' => [
+		//"Show" options
+		'show' => [
+			//Displays the page created datetime
+			'created' => TRUE,
+			//Displays the Shareaholic social buttons
+			//Remember you have to set app and site IDs. See `shareaholic.app_id` and `shareaholic.site_id`
+			'shareaholic' => FALSE			
+		]
+	],
 	//Posts
 	'post' => [
 		//"Show" options